multas_django/requirements.txt
Django>=4.2
pillow
reportlab
geopy
django-crispy-forms
crispy-bootstrap5
plotly
pandas
django-filter>=23.0
openpyxl>=3.1
